a huge upgrade

+++ repository/advertisements.py
@@ -7,7 +7,7 @@ from models import AdvertType, Advert
 from advert.models import AdvertIn
 from models import User
 from sqlalchemy import select
-
+from sqlalchemy.orm import selectinload
 
 class UserNotFound(Exception): ...
 
@@ -15,6 +15,8 @@ class UserNotFound(Exception): ...
 class AdvertNotFound(Exception): ...
 
 
+class CurrentUserError(Exception):
+    ...
 async def create_advert(
     db: AsyncSession,
     title: str,
@@ -41,10 +43,13 @@ async def create_advert(
     return AdvertIn
 
 
-async def delete_advert(db: AsyncSession, advert_id: int):
+async def delete_advert(db: AsyncSession, advert_id: int, current_user: User):
     try:
         result = await db.execute(select(Advert).where(Advert.id == advert_id))
         advert = result.scalars().one()
+
+        if advert.user_id != current_user.id:
+            raise CurrentUserError ("Данный пользователь не может удалить объявление!")
 
         await db.delete(advert)
         await db.commit()
@@ -56,3 +61,17 @@ async def get_list_advert(db: AsyncSession, limit_advert: int):
     result = await db.execute(select(Advert).limit(limit_advert))
     all_advert = result.scalars().all()
     return all_advert
+
+
+
+async def get_full_advert(db: AsyncSession, advert_id: int):
+    try:
+        result = await db.execute(
+            select(Advert)
+            .where(Advert.id == advert_id)
+            .options(selectinload(Advert.comments))
+        )
+        full_advert = result.scalars().one()
+    except NoResultFound as e:
+        raise AdvertNotFound("Объявления не существует!")
+    return full_advert
